+++ addon/components/fragment-plugin/fragment-insert.ts
@@ -4,9 +4,11 @@ import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 
 import { SayController } from '@lblod/ember-rdfa-editor';
+import { FragmentPluginConfig } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/fragment-plugin';
 
 interface Args {
   controller: SayController;
+  config: FragmentPluginConfig;
 }
 
 export default class FragmentInsertComponent extends Component<Args> {
@@ -14,6 +16,10 @@ export default class FragmentInsertComponent extends Component<Args> {
 
   get controller() {
     return this.args.controller;
+  }
+
+  get config() {
+    return this.args.config;
   }
 
   @action