+++ addon/components/fragment-plugin/search-modal.ts
@@ -1,16 +1,28 @@
 import { assert } from '@ember/debug';
 import { action } from '@ember/object';
+import { restartableTask, timeout } from 'ember-concurrency';
+import { task as trackedTask } from 'ember-resources/util/ember-concurrency';
 
 import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 
+import { fetchFragments } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/fragment-plugin/utils/fetch-data';
+import { FragmentPluginConfig } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/fragment-plugin';
+
 interface Args {
-  config: { endpoint: string };
+  config: FragmentPluginConfig;
   closeModal: () => void;
 }
 
 export default class FragmentPluginSearchModalComponent extends Component<Args> {
+  // Filtering
   @tracked inputSearchText: string | null = null;
+
+  // Pagination
+  @tracked totalCount = 0;
+
+  // Display
+  @tracked error: unknown;
 
   @action
   setInputSearchText(event: InputEvent) {
@@ -23,7 +35,36 @@ export default class FragmentPluginSearchModalComponent extends Component<Args> 
   }
 
   @action
-  closeModal() {
+  async closeModal() {
+    await this.fragmentsResource.cancel();
     this.args.closeModal();
   }
+
+  fragmentSearch = restartableTask(async () => {
+    await timeout(500);
+
+    const abortController = new AbortController();
+
+    try {
+      const queryResult = await fetchFragments({
+        endpoint: this.args.config.endpoint,
+        abortSignal: abortController.signal,
+        filter: {
+          name: this.inputSearchText ?? undefined,
+        },
+      });
+
+      this.totalCount = queryResult.totalCount;
+      return queryResult.results;
+    } catch (error) {
+      this.error = error;
+      return [];
+    } finally {
+      abortController.abort();
+    }
+  });
+
+  fragmentsResource = trackedTask(this, this.fragmentSearch, () => [
+    this.inputSearchText,
+  ]);
 }