+++ addon/plugins/fragment-plugin/index.ts
@@ -1,0 +1,3 @@
+export type FragmentPluginConfig = {
+  endpoint: string;
+};