+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Numbers inputted into a number variable are validated on defined min/max and if it is a number
 - Fragment insert modal
+- Query backend for fragments
 
 ### Fixed
 - Fixed woodpecker builds crashing on syntax errors